requirements.txt
torch
torchvision
timm
pillow
numpy
tqdm
pandas
joblib
flask
opencv-python
rapidfuzz
scikit-learn
paddleocr
sqlite3